write the CC lane policy once against transport traits
Nine method pairs existed once per transport as near-identical lambda adapters over the shared policy.
USB and DIN genuinely differ in four things: element type, how identity is read out of an element, how
many elements a message spans, and how a value byte is rewritten. Those become traits, and the policy is
written once.

The drain loops stay separate. USB fills a fixed-size transfer against a message-count allowance; DIN
paces bytes against a Q8 allowance and UART space and validates variable-length heads. Those are
different algorithms, and merging them would cost more than the duplication did.

MIDIQueueManager and MIDICCQueuePolicy move to the new midi_cc_policy.h so the traits can name them
without a header cycle. The scan the two pairs collapse into stops at the first element that is not a
channel CC rather than skipping it: classify_message routes only Continuous channel CCs to that lane, so
anything else means the lane is not aligned to the transport's span, and stepping on could hand the
scheduler a misaligned span to remove.

USB had no end-to-end drain coverage where DIN did, so add the two cases this change could have broken:
coalescing plus debt ordering, and Event CCs keeping their order and their duplicate values.
